Un-rename calculateHourlyScores
