+++ interfact-admin-dashboard/src/app/utils/calculateHourlyScores.ts
